:twisted_rightwards_arrows: Add feature reduce (#4)
* Setup devtools (#1)

* :wrench: Initialize yarn and npm

* :wrench: Add eslint and prettier

This commit adds eslint as linter and prettier as a code formatter. It also sets up a precommit hook to run js-files through eslint and prettier before committing.

* :wrench: Add flow

* :bug: Move flow-bin from dependencies to devDependencies

* :bug: Fix eslint extends

Fixes wrongly named library (airbnb) in .eslintrc to airbnb-base

* :wrench: Update .gitignore

Add build as a directory to ignore

* :wrench: Setup Rollup with Babel

* :art: Extract common settings

* :art: Move .babelrc to root

This commit will move .babelrc to the project root and also make use of the env prop to make it work in different processes.

* :white_check_mark: Add ava for testing

* :art: Strip flowtypes

Adds flow type support to babel and eslint

* :twisted_rightwards_arrows: Add module pipe (#2)

Adds the module pipe which takes any number of functions and creates a pipe – a larger function. Any argumnets passed to that function will then flow through all the smaller function, like in a pipe.

* :twisted_rightwards_arrows: Add module compose (#3)

This module takes, like pipe, any numer of functions and creates a reversed pipe
(from right to left) and creates a larger function where the next argument can
flow through.

* :wrench: Add global definitions to flow

* :spark: Add module reduce

* :bug: Fix issue with GenericFn

Got a warning if fn hade more than one arg

* :hammer: Refactore reduce to not use built in reduce

* :art: Use internal module reduce where possible